Update README with new name for app
I didn't realize that DingDong can be used as a derogatory term (https://www.theodysseyonline.com/5-truths-asian-stereotypes-american). Best to rename and be considerate of everyone. 

Heimdall is the new name. He's a character from Thor movies/comics (https://en.wikipedia.org/wiki/Heimdall_(comics)) and stands guard against enemies as well as provides passage to Asguardians. Seems like an apt name for an app that mostly will deal with accounts, users and auth to explore decomposition patterns.

The rest of the app needs to be renamed but this is a start.